pass dtype explicitly to sample function

+++ src/gluonts/distribution/binned.py
@@ -17,6 +17,7 @@ from typing import Tuple, List
 # Third-party imports
 import mxnet as mx
 from mxnet import gluon
+import numpy as np
 
 # First-party imports
 from gluonts.core.component import validated
@@ -157,7 +158,7 @@ class Binned(Distribution):
         a = centers_expanded.pick(idx, axis=-1)
         return a.swapaxes(0, 1)
 
-    def sample(self, num_samples=None):
+    def sample(self, num_samples=None, dtype=np.float32):
         def s(bin_probs):
             F = self.F
             indices = F.sample_multinomial(bin_probs)
